tests/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
